requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
Pillow>=9.0.0
bcrypt==4.0.1

# PostgreSQL dependencies
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Performance and monitoring
Flask-Compress==1.13
Flask-Limiter==3.5.0
redis==4.6.0

# Production server
gunicorn==21.2.0

# Development and testing
pytest==7.4.2
pytest-flask==1.3.0
black==23.9.1
flake8==6.1.0

# Security
cryptography==41.0.4

# Environment management
python-dotenv==1.0.0 

# Date utilities
python-dateutil==2.9.0
openpyxl==3.1.5
twilio==9.2.3